CleanUp CommonActivityUtils - move tint method to ThemeUtils
