+++ vector/src/main/java/im/vector/activity/PhoneNumberVerificationActivity.java
@@ -124,7 +124,7 @@ public class PhoneNumberVerificationActivity extends RiotAppCompatActivity imple
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.menu_phone_number_verification, menu);
-        CommonActivityUtils.tintMenuIcons(menu, ThemeUtils.getColor(this, R.attr.icon_tint_on_dark_action_bar_color));
+        ThemeUtils.tintMenuIcons(menu, ThemeUtils.getColor(this, R.attr.icon_tint_on_dark_action_bar_color));
         return true;
     }
