+++ vector/src/main/java/im/vector/adapters/VectorMessagesAdapter.java
@@ -378,7 +378,7 @@ public class VectorMessagesAdapter extends AbstractMessagesAdapter {
         mAlwaysShowTimeStamps = PreferencesManager.alwaysShowTimeStamps(VectorApp.getInstance());
         mHideReadReceipts = PreferencesManager.hideReadReceipts(VectorApp.getInstance());
 
-        mPadlockDrawable = CommonActivityUtils.tintDrawable(mContext,
+        mPadlockDrawable = ThemeUtils.tintDrawable(mContext,
                 ContextCompat.getDrawable(mContext, R.drawable.e2e_unencrypted), R.attr.settings_icon_tint_color);
     }
 
@@ -2459,7 +2459,7 @@ public class VectorMessagesAdapter extends AbstractMessagesAdapter {
         }
 
         Menu menu = popup.getMenu();
-        CommonActivityUtils.tintMenuIcons(menu, ThemeUtils.getColor(mContext, R.attr.settings_icon_tint_color));
+        ThemeUtils.tintMenuIcons(menu, ThemeUtils.getColor(mContext, R.attr.settings_icon_tint_color));
 
         // hide entries
         for (int i = 0; i < menu.size(); i++) {
